requirements.txt
customtkinter==5.2.2
Pillow==10.2.0
reportlab==4.1.0